TST: move _next_opt_len tests to fftpack
