+++ scipy/signal/tests/test_signaltools.py
@@ -17,7 +17,7 @@ from scipy import signal
 from scipy.signal import (
     correlate, convolve, convolve2d, fftconvolve, hann,
     hilbert, hilbert2, lfilter, lfilter_zi, filtfilt, butter, tf2zpk,
-    invres, invresz, vectorstrength, signaltools, lfiltic, tf2sos, sosfilt,
+    invres, invresz, vectorstrength, lfiltic, tf2sos, sosfilt,
     sosfilt_zi)
 from scipy.signal.signaltools import _filtfilt_gust
 
@@ -25,6 +25,7 @@ if sys.version_info.major >= 3 and sys.version_info.minor >= 5:
     from math import gcd
 else:
     from fractions import gcd
+
 
 class _TestConvolve(TestCase):
 
@@ -395,104 +396,6 @@ class TestFFTConvolve(TestCase):
             c = fftconvolve(a, b, 'full')
             d = np.convolve(a, b, 'full')
             assert_allclose(c, d, atol=1e-10, err_msg=msg)
-
-    def test_next_opt_len(self):
-        np.random.seed(1234)
-
-        def ns():
-            for j in range(1, 1000):
-                yield j
-            yield 2**5 * 3**5 * 4**5 + 1
-
-        for n in ns():
-            m = signaltools._next_opt_len(n)
-            msg = "n=%d, m=%d" % (n, m)
-
-            assert_(m >= n, msg)
-
-            # check regularity
-            k = m
-            for d in [2, 3, 5]:
-                while True:
-                    a, b = divmod(k, d)
-                    if b == 0:
-                        k = a
-                    else:
-                        break
-            assert_equal(k, 1, err_msg=msg)
-
-    def test_next_opt_len_strict(self):
-        hams = {
-            1: 1, 2: 2, 3: 3, 4: 4, 5: 5, 6: 6, 7: 8, 8: 8, 14: 15, 15: 15,
-            16: 16, 17: 18, 1021: 1024, 1536: 1536, 51200000: 51200000,
-            510183360: 510183360, 510183360 + 1: 512000000,
-            511000000: 512000000,
-            854296875: 854296875, 854296875 + 1: 859963392,
-            196608000000: 196608000000, 196608000000 + 1: 196830000000,
-            8789062500000: 8789062500000, 8789062500000 + 1: 8796093022208,
-            206391214080000: 206391214080000,
-            206391214080000 + 1: 206624260800000,
-            470184984576000: 470184984576000,
-            470184984576000 + 1: 470715894135000,
-            7222041363087360: 7222041363087360,
-            7222041363087360 + 1: 7230196133913600,
-            # power of 5    5**23
-            11920928955078125: 11920928955078125,
-            11920928955078125 - 1: 11920928955078125,
-            # power of 3    3**34
-            16677181699666569: 16677181699666569,
-            16677181699666569 - 1: 16677181699666569,
-            # power of 2   2**54
-            18014398509481984: 18014398509481984,
-            18014398509481984 - 1: 18014398509481984,
-            # above this, int(ceil(n)) == int(ceil(n+1))
-            19200000000000000: 19200000000000000,
-            19200000000000000 + 1: 19221679687500000,
-            288230376151711744: 288230376151711744,
-            288230376151711744 + 1: 288325195312500000,
-            288325195312500000 - 1: 288325195312500000,
-            288325195312500000: 288325195312500000,
-            288325195312500000 + 1: 288555831593533440,
-            # power of 3    3**83
-            3990838394187339929534246675572349035227 - 1:
-                3990838394187339929534246675572349035227,
-            3990838394187339929534246675572349035227:
-                3990838394187339929534246675572349035227,
-            # power of 2     2**135
-            43556142965880123323311949751266331066368 - 1:
-                43556142965880123323311949751266331066368,
-            43556142965880123323311949751266331066368:
-                43556142965880123323311949751266331066368,
-            # power of 5      5**57
-            6938893903907228377647697925567626953125 - 1:
-                6938893903907228377647697925567626953125,
-            6938893903907228377647697925567626953125:
-                6938893903907228377647697925567626953125,
-            # http://www.drdobbs.com/228700538
-            # 2**96 * 3**1 * 5**13
-            290142196707511001929482240000000000000 - 1:
-                290142196707511001929482240000000000000,
-            290142196707511001929482240000000000000:
-                290142196707511001929482240000000000000,
-            290142196707511001929482240000000000000 + 1:
-                290237644800000000000000000000000000000,
-            # 2**36 * 3**69 * 5**7
-            4479571262811807241115438439905203543080960000000 - 1:
-                4479571262811807241115438439905203543080960000000,
-            4479571262811807241115438439905203543080960000000:
-                4479571262811807241115438439905203543080960000000,
-            4479571262811807241115438439905203543080960000000 + 1:
-                4480327901140333639941336854183943340032000000000,
-            # 2**37 * 3**44 * 5**42
-            30774090693237851027531250000000000000000000000000000000000000 - 1:
-                30774090693237851027531250000000000000000000000000000000000000,
-            30774090693237851027531250000000000000000000000000000000000000:
-                30774090693237851027531250000000000000000000000000000000000000,
-            30774090693237851027531250000000000000000000000000000000000000 + 1:
-                30778180617309082445871527002041377406962596539492679680000000,
-        }
-        for x, y in hams.items():
-            assert_equal(signaltools._next_opt_len(x), y)
 
     def test_invalid_shapes(self):
         # By "invalid," we mean that no one
